feat: add a GitHub check that prevents non-master non-backports
When a PR is raised against a non-master branch Trop will try to figure out which
PR it is "backporting".  If it can't determine it Trop will mark that PR as
"failed" with a GitHub Check.

This ensures that no change can land in a release branch without having a
corresponding commit / PR in master.  In cases where this must be overridden
an Admin can bypass the merge check.